Added initial pass over the rules
All files are in js files and are working when the folder is called tslint.json.

Still need to go over some of the todo ones and get the 2 missing rules in somewhere.

Next step is first to get full instructions on how to integrate this (probably with git submodule) into another project.